// src/generated/live/LiveUiState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LiveNote } from "./LiveNote";
import type { PlayerIdentityLink } from "./PlayerIdentityLink";
import type { PlayerProfile } from "./PlayerProfile";
import type { PremadeMark } from "./PremadeMark";
import type { SavedAccount } from "./SavedAccount";

export type LiveUiState = { myAccounts: Array<SavedAccount>, notes: Array<LiveNote>, reviewed: Array<string>, theme: string | null, players: Array<PlayerProfile>, premades: Array<PremadeMark>, identityLinks: Array<PlayerIdentityLink>, };